+++ src/app/hooks/queries/useAuthMe.ts
@@ -1,29 +1,6 @@
-import { env } from '@/env';
+import { fetchAuthMe } from '@/app/lib/fetchAuthMe';
 import type { User } from '@prisma/client';
-import {
-  type UseSuspenseQueryResult,
-  isServer,
-  queryOptions,
-  useSuspenseQuery,
-} from '@tanstack/react-query';
-
-const BASE_URL = env.NEXT_PUBLIC_API_BASE_URL;
-
-export const fetchAuthMe = async (): Promise<User> => {
-  const headers = {
-    ...(isServer ? { Cookie: (await (await import('next/headers')).cookies()).toString() } : {}),
-  };
-
-  const response = await fetch(`${BASE_URL}/api/auth/me`, {
-    headers,
-  });
-
-  if (!response.ok) {
-    throw new Error(`${response.status}`);
-  }
-
-  return response.json();
-};
+import { type UseSuspenseQueryResult, queryOptions, useSuspenseQuery } from '@tanstack/react-query';
 
 const useAuthMeOptions = () => {
   return queryOptions<User>({
